+++ drf_spectacular/contrib/rest_framework_simplejwt.py
@@ -1,3 +1,4 @@
+from drf_spectacular.drainage import warn
 from drf_spectacular.extensions import OpenApiAuthenticationExtension
 
 
@@ -7,8 +8,6 @@ class SimpleJWTScheme(OpenApiAuthenticationExtension):
 
     def get_security_definition(self, auto_schema):
         from rest_framework_simplejwt.settings import api_settings
-
-        from drf_spectacular.plumbing import warn
 
         if len(api_settings.AUTH_HEADER_TYPES) > 1:
             warn(
