+++ drf_spectacular/management/commands/spectacular.py
@@ -4,7 +4,7 @@ from django.core.management.base import BaseCommand
 from django.utils import translation
 from django.utils.module_loading import import_string
 
-from drf_spectacular.plumbing import GENERATOR_STATS
+from drf_spectacular.drainage import GENERATOR_STATS
 from drf_spectacular.renderers import OpenApiJsonRenderer, OpenApiYamlRenderer
 from drf_spectacular.settings import spectacular_settings
 from drf_spectacular.validation import validate_schema
